pyproject.toml
[project]
name = "self-heal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "google-adk>=1.19.0",
]
